@@ -14,7 +14,7 @@ description: bsteele Music App in Flutter
 version: 0.0.1+1
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=2.12.0-179.0.dev <3.0.0"
 
 dependencies:
   flutter:
